HTML Result: move rendering code from core to the plugin itself
This change removes the knowledge, and dependency, that the current
Avocado core has of the HTML rendering code.

The only user visible change is that the UI message about the
generated HTML file now appears one line after it used to appear.  The
reason is not because of the HTML plugin itself, but because the
results are now generated outside the context of the job, and that is,
chronologically a bit later than it used to be.
