+++ docs/source/Replay.rst
@@ -20,8 +20,8 @@ Let's see an example. First, running a simple job with two urls::
       (1/2) /bin/true: PASS (0.01 s)
       (2/2) /bin/false: FAIL (0.01 s)
      RESULTS    : PASS 1 | ERROR 0 | FAIL 1 | SKIP 0 | WARN 0 | INTERRUPT 0
-     JOB HTML   : $HOME/avocado/job-results/job-2016-01-11T16.14-825b860/html/results.html
      TESTS TIME : 0.02 s
+     JOB HTML   : $HOME/avocado/job-results/job-2016-01-11T16.14-825b860/html/results.html
 
 Now we can replay the job by running::
 
@@ -33,8 +33,8 @@ Now we can replay the job by running::
       (1/2) /bin/true: PASS (0.01 s)
       (2/2) /bin/false: FAIL (0.01 s)
      RESULTS    : PASS 1 | ERROR 0 | FAIL 1 | SKIP 0 | WARN 0 | INTERRUPT 0
-     JOB HTML   : $HOME/avocado/job-results/job-2016-01-11T16.18-55a0d10/html/results.html
      TESTS TIME : 0.01 s
+     JOB HTML   : $HOME/avocado/job-results/job-2016-01-11T16.18-55a0d10/html/results.html
 
 The replay feature will retrieve the original job urls, the multiplex
 tree and the configuration. Let's see another example, now using
@@ -93,8 +93,8 @@ multiplex file::
       (47/48) /bin/false;23: FAIL (0.01 s)
       (48/48) /bin/false;24: FAIL (0.01 s)
      RESULTS    : PASS 24 | ERROR 0 | FAIL 24 | SKIP 0 | WARN 0 | INTERRUPT 0
-     JOB HTML   : $HOME/avocado/job-results/job-2016-01-11T21.56-bd6aa3b/html/results.html
      TESTS TIME : 0.29 s
+     JOB HTML   : $HOME/avocado/job-results/job-2016-01-11T21.56-bd6aa3b/html/results.html
 
 We can replay the job as is, using ``$ avocado run --replay latest``,
 or replay the job ignoring the multiplex file, as below::
@@ -108,8 +108,8 @@ or replay the job ignoring the multiplex file, as below::
       (1/2) /bin/true: PASS (0.01 s)
       (2/2) /bin/false: FAIL (0.01 s)
      RESULTS    : PASS 1 | ERROR 0 | FAIL 1 | SKIP 0 | WARN 0 | INTERRUPT 0
-     JOB HTML   : $HOME/avocado/job-results/job-2016-01-11T22.01-d5a4618/html/results.html
      TESTS TIME : 0.02 s
+     JOB HTML   : $HOME/avocado/job-results/job-2016-01-11T22.01-d5a4618/html/results.html
 
 Also, it is possible to replay only the variants that faced a given
 result, using the option ``--replay-test-status``. See the example below::
@@ -168,8 +168,8 @@ result, using the option ``--replay-test-status``. See the example below::
      (47/48) /bin/false;23: FAIL (0.01 s)
      (48/48) /bin/false;24: FAIL (0.01 s)
     RESULTS    : PASS 0 | ERROR 0 | FAIL 24 | SKIP 24 | WARN 0 | INTERRUPT 0
-    JOB HTML   : $HOME/avocado/job-results/job-2016-01-12T00.38-2e1dc41/html/results.html
     TESTS TIME : 0.19 s
+    JOB HTML   : $HOME/avocado/job-results/job-2016-01-12T00.38-2e1dc41/html/results.html
 
 When replaying jobs that were executed with the ``--failfast on`` option, you
 can disable the ``failfast`` option using ``--failfast off`` in the replay job.
@@ -186,8 +186,8 @@ below::
      TESTS      : 1
       (1/1) /bin/true: PASS (0.01 s)
      RESULTS    : PASS 1 | ERROR 0 | FAIL 0 | SKIP 0 | WARN 0 | INTERRUPT 0
-     JOB HTML   : /tmp/avocado_results/job-2016-01-11T22.10-f1b1c87/html/results.html
      TESTS TIME : 0.01 s
+     JOB HTML   : /tmp/avocado_results/job-2016-01-11T22.10-f1b1c87/html/results.html
 
 Trying to replay the job, it fails::
 
@@ -203,5 +203,5 @@ In this case, we have to inform where the job results directory is located::
      TESTS      : 1
       (1/1) /bin/true: PASS (0.01 s)
      RESULTS    : PASS 1 | ERROR 0 | FAIL 0 | SKIP 0 | WARN 0 | INTERRUPT 0
-     JOB HTML   : $HOME/avocado/job-results/job-2016-01-11T22.15-19c76ab/html/results.html
      TESTS TIME : 0.01 s
+     JOB HTML   : $HOME/avocado/job-results/job-2016-01-11T22.15-19c76ab/html/results.html
